feat(INTERNAL-1356): new firing form

+++ src/modules/userMethods.ts
@@ -6,7 +6,6 @@ import { SessionUser } from '../utils/auth';
 import { suggestionsTake } from '../utils/suggestions';
 import { defaultTake } from '../utils';
 import { getCorporateEmail } from '../utils/getCorporateEmail';
-import { getLastSupplementalPositions } from '../utils/supplementalPositions';
 import { calculateDiffBetweenArrays } from '../utils/calculateDiffBetweenArrays';
 
 import {
@@ -174,6 +173,7 @@ export const userMethods = {
 
     removeFromGroup: async (data: RemoveUserFromGroup) => {
         const membership = await prisma.membership.findUnique({ where: { userId_groupId: data } });
+
         if (membership?.archived) {
             throw new TRPCError({ code: 'BAD_REQUEST', message: tr('Cannot edit archived membership') });
         }
@@ -458,7 +458,7 @@ export const userMethods = {
 
         const workEndDate = new Date();
 
-        if (data.active === false) {
+        if (data.active === false && !data.scheduled) {
             await prisma.membership.deleteMany({ where: { userId: data.id, group: { organizational: true } } });
 
             await prisma.supplementalPosition.updateMany({
@@ -477,21 +477,8 @@ export const userMethods = {
                     workEndDate,
                 },
             });
-        } else {
-            const { positions: selectedPosition } = getLastSupplementalPositions(positions);
-
-            await prisma.supplementalPosition.updateMany({
-                where: {
-                    id: {
-                        in: selectedPosition.map((p) => p.id),
-                    },
-                },
-                data: {
-                    status: 'ACTIVE',
-                    workEndDate: null,
-                },
-            });
         }
+
         await prisma.membership.updateMany({ where: { userId: data.id }, data: { archived: !data.active } });
         return prisma.user.update({
             where: { id: data.id },
@@ -598,13 +585,18 @@ export const userMethods = {
 
     getMailingList: async (
         mailingType: MailingSettingType,
-        organizationUnitId: string,
+        organizationUnitIds: string[],
         additionUsersIds?: string[],
     ) => {
         const mailingList = await prisma.user.findMany({
             where: {
                 OR: [
-                    { mailingSettings: { some: { [mailingType]: true, organizationUnitId } }, active: true },
+                    {
+                        mailingSettings: {
+                            some: { [mailingType]: true, organizationUnitId: { in: organizationUnitIds } },
+                        },
+                        active: true,
+                    },
                     { id: { in: additionUsersIds } },
                 ],
             },
@@ -613,7 +605,7 @@ export const userMethods = {
 
         const additionalEmails = await mailSettingsMethods.getAdditionEmails({
             mailingType,
-            organizationUnitId,
+            organizationUnitIds,
         });
 
         const users = [
